+++ src/config.rs
@@ -110,6 +110,17 @@ fn yamlloader(args: &HashMap<String, tera::Value>) -> tera::Result<tera::Value> 
     serde_yaml::from_reader(f).map_err(|_| "Failed to read file".into())
 }
 
+fn jsonloader(args: &HashMap<String, tera::Value>) -> tera::Result<tera::Value> {
+    let file = match args.get("file") {
+        Some(file) => file,
+        None => return Err("file parameter is required".into()),
+    };
+
+    let path = from_value::<String>(file.clone())?;
+    let f = std::fs::File::open(path)?;
+    serde_json::from_reader(f).map_err(|_| "Failed to read file".into())
+}
+
 // Renders a template from a given string.
 pub fn render_template(template: &str, source_filename: &Path) -> Result<String, Error> {
     let template_path = source_filename.display().to_string();
@@ -119,8 +130,9 @@ pub fn render_template(template: &str, source_filename: &Path) -> Result<String,
     // Read the template using tera
     let mut tera = Tera::default();
 
-    // Allow templates to load yaml files as Values.
+    // Allow templates to load yaml and json files as Values.
     tera.register_function("yamlload", yamlloader);
+    tera.register_function("jsonload", jsonloader);
 
     tera.add_raw_template(&template_path, template)
         .map_err(|e| errors::FlokiError::ProblemRenderingTemplate {
@@ -209,9 +221,6 @@ fn default_entrypoint() -> Entrypoint {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::image::Image;
-    use std::io::Write;
-    use tempfile::NamedTempFile;
 
     #[derive(Debug, PartialEq, Serialize, Deserialize)]
     struct TestShellConfig {
@@ -307,6 +316,14 @@ mod test {
     #[test]
     fn test_tera_yamlload() -> Result<(), Box<dyn std::error::Error>> {
         let template = r#"{% set values = yamlload(file="test_resources/values.yaml") %}shell: {{ values.foo }}"#;
+        let config = render_template(template, Path::new("floki"))?;
+        assert_eq!(config, "shell: bar");
+        Ok(())
+    }
+
+    #[test]
+    fn test_tera_jsonload() -> Result<(), Box<dyn std::error::Error>> {
+        let template = r#"{% set values = jsonload(file="test_resources/values.json") %}shell: {{ values.foo }}"#;
         let config = render_template(template, Path::new("floki"))?;
         assert_eq!(config, "shell: bar");
         Ok(())